H264_nal: kill a warning

+++ modules/codec/omxil/omxil.h
@@ -102,7 +102,7 @@ struct decoder_sys_t
 
     date_t end_date;
 
-    int i_nal_size_length; /* Length of the NAL size field for H264 */
+    size_t i_nal_size_length; /* Length of the NAL size field for H264 */
     int b_use_pts;
 
 };
